Update for CosmologyHalos.jl/halo.jl refactor
Fixed code formatting and added tests for the NFW type, which is now a dependency of CosmologyHalos.jl / halo.jl after the refactor. Might be good to add tests for units to NFW, but probably not entirely necessary as the Unitful behavior should be covered by the tests with the isothermal and exponential disk profiles.